+++ packages/components/card-tabs/src/CardTab/CardTabActions/CardTabActions.tsx
@@ -3,17 +3,11 @@ import Cruds from '@synerise/ds-cruds';
 import * as S from '../CardTab.styles';
 import { Props } from './CardTabActions.types';
 
-const CardTabActions: React.FC<Props> = ({
-  changeNameAvailable,
-  enterEditNameMode,
-  onDuplicateTab,
-  onRemoveTab,
-  texts,
-}) => {
+const CardTabActions: React.FC<Props> = ({ onChangeName, enterEditNameMode, onDuplicateTab, onRemoveTab, texts }) => {
   return (
     <S.CardTabSuffix data-testid="card-tab-suffix">
       <Cruds
-        onEdit={changeNameAvailable ? enterEditNameMode : undefined}
+        onEdit={onChangeName ?? enterEditNameMode}
         editTooltip={texts?.changeNameTooltip}
         onDuplicate={onDuplicateTab}
         duplicateTooltip={texts?.duplicateTooltip}