+++ packages/components/factors/src/FactorValue/FactorValue.tsx
@@ -20,13 +20,13 @@ const FactorValue: React.FC<FactorValueProps> = ({
   preventAutoloadData,
   getPopupContainerOverride,
   onActivate,
+  onDeactivate,
+  error,
 }) => {
   const inputType = React.useMemo(() => {
-    const InputComponent = selectedFactor.input;
+    const InputComponent: React.ElementType = selectedFactor.input;
 
     return (
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
       <InputComponent
         key={`${factorKey}-${selectedFactor.name}`}
         value={value}
@@ -45,6 +45,8 @@ const FactorValue: React.FC<FactorValueProps> = ({
         preventAutoloadData={preventAutoloadData}
         getPopupContainerOverride={getPopupContainerOverride}
         onActivate={onActivate}
+        onDeactivate={onDeactivate}
+        error={error}
       />
     );
   }, [
@@ -64,8 +66,10 @@ const FactorValue: React.FC<FactorValueProps> = ({
     loading,
     onParamsClick,
     preventAutoloadData,
-    onActivate,
     getPopupContainerOverride,
+    onActivate,
+    onDeactivate,
+    error,
   ]);
 
   return <S.FactorInput withoutTypeSelector={withoutTypeSelector}>{inputType}</S.FactorInput>;