+++ packages/components/subject/src/Subject.tsx
@@ -16,7 +16,9 @@ const Subject: React.FC<SubjectProps> = ({
   type = 'parameter',
   getPopupContainerOverride,
   onActivate,
+  onDeactivate,
   texts,
+  opened,
 }) => {
   const [dropdownVisible, setDropdownVisible] = React.useState(false);
 
@@ -24,9 +26,17 @@ const Subject: React.FC<SubjectProps> = ({
     return type && ['event', 'context'].indexOf(type) >= 0 ? 'cyan' : 'green';
   }, [type]);
 
-  const onDropdownVisibilityChange = React.useCallback((value: boolean) => value && onActivate && onActivate(), [
-    onActivate,
-  ]);
+  const onDropdownVisibilityChange = React.useCallback(
+    (value: boolean) => {
+      value && onActivate && onActivate('');
+      !value && onDeactivate && onDeactivate();
+    },
+    [onActivate, onDeactivate]
+  );
+
+  React.useEffect(() => {
+    setDropdownVisible(Boolean(opened));
+  }, [opened]);
 
   return (
     <S.Subject>