+++ packages/components/table/src/VirtualTable/VirtualTableRow.tsx
@@ -1,11 +1,13 @@
 import * as React from 'react';
 import classNames from 'classnames';
+import { areEqual } from 'react-window';
 import InfiniteLoaderItem from '../InfiniteScroll/InfiniteLoaderItem';
 import { InfiniteScrollProps } from '../InfiniteScroll/constants';
 import { RowSelection, DSColumnType, DSTableProps } from '../Table.types';
 import { EXPANDED_ROW_PROPERTY } from './VirtualTable';
 import * as S from './VirtualTable.styles';
-import { calculatePixels } from '../utils/calculatePixels';
+import { RowStar } from '../hooks/useRowStar';
+import { getValueFromPath, calculatePixels } from '../utils';
 
 interface Props<T> {
   data: {
@@ -15,6 +17,7 @@ interface Props<T> {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     mergedColumns: any[];
     selection?: RowSelection<T>;
+    rowStar?: RowStar<T>;
     onRowClick?: (row: T) => void;
     defaultTableProps?: DSTableProps<T>;
   };
@@ -27,58 +30,83 @@ const isColumnSortingActive = <T extends unknown>(columns: DSColumnType<T>[], co
 
 const calculateToPixelsIfDefined = (value: string | number | undefined | null): number | undefined | null =>
   value ? calculatePixels(value) : (value as number);
-class VirtualTableRow<T> extends React.PureComponent<Props<T>> {
-  render(): React.ReactNode {
-    const { index, style, data } = this.props;
-    const { mergedColumns, onRowClick, selection, dataSource, cellHeight, infiniteScroll, defaultTableProps } = data;
-    const rowData = dataSource[index];
 
+function VirtualTableRow<T extends object>({
+  index,
+  style,
+  data: { mergedColumns, onRowClick, selection, rowStar, dataSource, cellHeight, infiniteScroll, defaultTableProps },
+}: Props<T>): React.ReactElement {
+  const renderColumn = React.useCallback(
+    (column: DSColumnType<T>, rowData: T, columnIndex: number): React.ReactNode => {
+      if (rowData[EXPANDED_ROW_PROPERTY] && column.childRender) {
+        return column.childRender(getValueFromPath(rowData, column.dataIndex), rowData, columnIndex);
+      }
+      return column.render
+        ? column.render(getValueFromPath(rowData, column.dataIndex), rowData, columnIndex)
+        : getValueFromPath(rowData, column.dataIndex);
+    },
+    []
+  );
+
+  const rowData = React.useMemo(() => dataSource[index], [dataSource, index]);
+
+  const infiniteLoader = React.useMemo(() => {
     return (
-      // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions
-      <>
+      infiniteScroll &&
+      index === dataSource.length - 1 && (
         <S.RowWrapper
-          className={classNames('virtual-table-row', {
-            'ds-expanded-row': rowData[EXPANDED_ROW_PROPERTY],
-          })}
-          style={style}
-          onClick={(): void => onRowClick && onRowClick(rowData)}
+          style={{ ...style, top: `${Number(style.top) + cellHeight}px`, height: '64px', padding: '16px 24px' }}
         >
-          {mergedColumns.map((column, columnIndex) => {
-            return (
-              <S.ColWrapper
-                className={classNames(
-                  'virtual-table-cell',
-                  {
-                    'virtual-table-cell-last': columnIndex === mergedColumns.length - 1,
-                    'ant-table-selection-column': columnIndex === 0 && selection,
-                    'ds-expanded-row-first': rowData[EXPANDED_ROW_PROPERTY] && columnIndex === 0,
-                    'ds-expanded-row-data':
-                      rowData[EXPANDED_ROW_PROPERTY] &&
-                      ((columnIndex === 1 && selection) || (columnIndex === 0 && !selection)),
-                  },
-                  isColumnSortingActive<T>(defaultTableProps?.columns || [], column) && 'ant-table-column-sort',
-                  column.className
-                )}
-                key={`row-${index}-column-${column.dataIndex || column.key}`}
-                minWidth={calculateToPixelsIfDefined(column?.minWidth)}
-                width={column.width}
-                maxWidth={calculateToPixelsIfDefined(column?.maxWidth)}
-              >
-                {column.render ? column.render(rowData[column.dataIndex], rowData) : rowData[column.dataIndex]}
-              </S.ColWrapper>
-            );
-          })}
+          <InfiniteLoaderItem infiniteScroll={infiniteScroll} />
         </S.RowWrapper>
-        {infiniteScroll && index === dataSource.length - 1 && (
-          <S.RowWrapper
-            style={{ ...style, top: `${Number(style.top) + cellHeight}px`, height: '64px', padding: '16px 24px' }}
-          >
-            <InfiniteLoaderItem infiniteScroll={infiniteScroll} />
-          </S.RowWrapper>
-        )}
-      </>
+      )
     );
-  }
+  }, [cellHeight, dataSource.length, index, infiniteScroll, style]);
+
+  return (
+    <>
+      <S.RowWrapper
+        className={classNames('virtual-table-row', {
+          'ds-expanded-row': rowData[EXPANDED_ROW_PROPERTY],
+        })}
+        style={style}
+        onClick={(event): void => {
+          event.stopPropagation();
+          onRowClick && onRowClick(rowData);
+        }}
+      >
+        {mergedColumns.map((column, columnIndex) => {
+          const firstWithSelectionAndStar = selection && rowStar && columnIndex === 2;
+          const firstWithSelectionOrStar = (selection || rowStar) && columnIndex === 1;
+          const firstWithoutSelectionAndStar = columnIndex === 0 && !selection && !rowStar;
+          return (
+            <S.ColWrapper
+              className={classNames(
+                'virtual-table-cell',
+                {
+                  'virtual-table-cell-last': columnIndex === mergedColumns.length - 1,
+                  'ant-table-selection-column': columnIndex === 0 && selection,
+                  'ds-expanded-row-first': rowData[EXPANDED_ROW_PROPERTY] && columnIndex === 0,
+                  'ds-expanded-row-data':
+                    rowData[EXPANDED_ROW_PROPERTY] &&
+                    (firstWithoutSelectionAndStar || firstWithSelectionOrStar || firstWithSelectionAndStar),
+                },
+                isColumnSortingActive<T>(defaultTableProps?.columns || [], column) && 'ant-table-column-sort',
+                column.className
+              )}
+              key={`row-${index}-column-${column.dataIndex || column.key}`}
+              minWidth={calculateToPixelsIfDefined(column?.minWidth)}
+              width={column.width}
+              maxWidth={calculateToPixelsIfDefined(column?.maxWidth)}
+            >
+              {renderColumn(column, rowData, columnIndex)}
+            </S.ColWrapper>
+          );
+        })}
+      </S.RowWrapper>
+      {infiniteLoader}
+    </>
+  );
 }
 
-export default VirtualTableRow;
+export default React.memo(VirtualTableRow, areEqual);