+++ packages/components/page-header/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.12](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.11...@synerise/ds-page-header@0.14.12) (2022-04-29)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.11](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.10...@synerise/ds-page-header@0.14.11) (2022-04-08)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.10](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.9...@synerise/ds-page-header@0.14.10) (2022-04-05)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.9](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.8...@synerise/ds-page-header@0.14.9) (2022-03-30)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.8](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.7...@synerise/ds-page-header@0.14.8) (2022-03-24)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.7](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.6...@synerise/ds-page-header@0.14.7) (2022-03-24)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.6](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.5...@synerise/ds-page-header@0.14.6) (2022-03-21)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.5](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.4...@synerise/ds-page-header@0.14.5) (2022-03-14)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
+## [0.14.4](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.3...@synerise/ds-page-header@0.14.4) (2022-02-25)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
+
+
+
+
 ## [0.14.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.14.2...@synerise/ds-page-header@0.14.3) (2022-02-16)
 
 **Note:** Version bump only for package @synerise/ds-page-header