+++ packages/components/progress-bar/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.6.9...@synerise/ds-progress-bar@0.6.10) (2022-04-29)
+
+**Note:** Version bump only for package @synerise/ds-progress-bar
+
+
+
+
+
+## [0.6.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.6.8...@synerise/ds-progress-bar@0.6.9) (2022-04-05)
+
+**Note:** Version bump only for package @synerise/ds-progress-bar
+
+
+
+
+
+## [0.6.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.6.7...@synerise/ds-progress-bar@0.6.8) (2022-03-24)
+
+**Note:** Version bump only for package @synerise/ds-progress-bar
+
+
+
+
+
+## [0.6.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.6.6...@synerise/ds-progress-bar@0.6.7) (2022-03-24)
+
+**Note:** Version bump only for package @synerise/ds-progress-bar
+
+
+
+
+
+## [0.6.6](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.6.5...@synerise/ds-progress-bar@0.6.6) (2022-03-14)
+
+**Note:** Version bump only for package @synerise/ds-progress-bar
+
+
+
+
+
 ## [0.6.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.6.4...@synerise/ds-progress-bar@0.6.5) (2021-12-29)
 
 **Note:** Version bump only for package @synerise/ds-progress-bar